requirements.txt
django==3.1.1
djangorestframework==3.9.3